python: Move the docs for dazl.ledger out of source code and into a dedicated rst file.
